# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartFlow Next",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/story-drafts/{draft_id}/assets": {
      "get": {
        "summary": "List Assets",
        "operationId": "list_assets_api_story_drafts__draft_id__assets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetResponse"
                  },
                  "title": "Response List Assets Api Story Drafts  Draft Id  Assets Get"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND or ROUTE_NOT_FOUND",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-permission": "stories:drafts:read"
      },
      "post": {
        "summary": "Import Asset",
        "operationId": "import_asset_api_story_drafts__draft_id__assets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "field",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 40,
              "title": "Field"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 80,
              "pattern": "^[a-zA-Z0-9_-]+$",
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "x-file-name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 2000,
              "title": "X-File-Name"
            }
          },
          {
            "name": "x-file-size",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 524288000,
              "exclusiveMinimum": 0,
              "title": "X-File-Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND or ROUTE_NOT_FOUND",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-permission": "stories:drafts:write"
      }
    },
    "/api/story-drafts/{draft_id}/assets/{asset_id}": {
      "get": {
        "summary": "Download Asset",
        "operationId": "download_asset_api_story_drafts__draft_id__assets__asset_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND or ROUTE_NOT_FOUND",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-permission": "stories:drafts:read"
      }
    },
    "/api/story-drafts": {
      "post": {
        "summary": "Create",
        "operationId": "create_api_story_drafts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 80,
              "pattern": "^[a-zA-Z0-9_-]+$",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND or ROUTE_NOT_FOUND",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-permission": "stories:drafts:write"
      },
      "get": {
        "summary": "List Drafts",
        "operationId": "list_drafts_api_story_drafts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DraftSummary"
                  },
                  "title": "Response List Drafts Api Story Drafts Get"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND or ROUTE_NOT_FOUND",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-permission": "stories:drafts:read"
      }
    },
...
    "/api/diagnostics/drafts": {
      "get": {
        "summary": "Audit",
        "operationId": "audit_api_diagnostics_drafts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DraftAudit"
                  },
                  "title": "Response Audit Api Diagnostics Drafts Get"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND or ROUTE_NOT_FOUND",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-permission": "diagnostics:read"
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "field": {
            "type": "string",
            "title": "Field"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "missing": {
            "type": "boolean",
            "title": "Missing"
          }
        },
        "type": "object",
        "required": [
          "id",
          "field",
          "name",
          "size",
          "sha256",
          "missing"
        ],
        "title": "AssetResponse"
      },
...
      "DraftAudit": {
        "properties": {
          "draft_id": {
            "type": "string",
            "title": "Draft Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "active_step": {
            "type": "integer",
            "title": "Active Step"
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "draft_id",
          "revision",
          "active_step",
          "updated_at"
        ],
        "title": "DraftAudit"
      },
...
      "DraftInput": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "active_step": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Active Step",
            "default": 0
          },
          "config": {
            "$ref": "#/components/schemas/DraftConfig"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DraftInput"
      },
...
      "DraftResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "active_step": {
            "type": "integer",
            "title": "Active Step"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At"
          },
          "config": {
            "$ref": "#/components/schemas/DraftConfig"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/DraftIssue"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "revision",
          "schema_version",
          "active_step",
          "created_at",
          "updated_at",
          "config",
          "issues"
        ],
        "title": "DraftResponse"
      },
      "DraftSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "active_step": {
            "type": "integer",
            "title": "Active Step"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "revision",
          "schema_version",
          "active_step",
          "created_at",
          "updated_at"
        ],
        "title": "DraftSummary"
      },
...
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "recovery": {
            "type": "string",
            "title": "Recovery"
          },
          "http_status": {
            "type": "integer",
            "title": "Http Status"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "prepare",
                  "generate",
                  "save",
                  "done"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "recovery",
          "http_status",
          "trace_id",
          "stage"
        ],
        "title": "ErrorInfo"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/InvalidField"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "InvalidField": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "type"
        ],
        "title": "InvalidField"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}